[MS-21] Refactor: 애플로그인 디렉터리 구조 변경 및 에러 핸들링
